requirements.txt
asgiref
boto3
botocore
certifi
cffi
charset-normalizer
cryptography
dj-database-url
Django==4.2.7
django-anymail
django-cors-headers
django-jazzmin
django-storages
django-ckeditor-5
djangorestframework
djangorestframework-simplejwt
drf-yasg
environs
gunicorn
idna
inflection
jmespath
marshmallow
packaging
psycopg2
pycparser
PyJWT
python-dateutil
python-dotenv
pytz==2023.3.post1
PyYAML
requests
s3transfer
shortuuid
six
sqlparse
stripe
typing_extensions
tzdata
uritemplate
urllib3
